Optimize buffer when no over-read
If we read exactly N bytes, we can cheaply reset the buffer back to 0.
The more we keep the buffer close to the start, the less we're likely
to need to wrap.

Reworked test fragmentation so that it not only fragments into smaller
messages but it also fragments into larger ones.