Reorganize initial client setup
The handshake_buffer is now valid for the minimium amount of time necessary,
with the memory freed as soon as the handshake reply is sent (or the connection
is rejected).

H.init now takes a full Handshake object, which now contains the raw headers.
The library no longer validates the websocket path (and so websocket.Config.path
is no longer an option). H.init can check the handhake's method/path if needed.