Encapsluate buffers, refactor fragmentation handling
Introduced a Buffer that wraps an []u8 with a type, where type can be static,
dynamic or pooled (future). The goal is to be able to make smarter decisions
based on the type, e.g. fragmented messages don't need to clone dynamic or
pooled payloads, they just need to take ownership of it.

Reworked how the fragmenter works, to simplify it and give it a cleaner
lifecycle so that it can "own" buffers.