NonBlocking worker buffer strategies.
App can either opt to have a per-connection buffer available at all time, or
use a buffer pool to allocate buffers just-in-time.

This allows the memory usage to be optimized based on how clients are expected
to interact with the server. When all clients are expected to send a steady
stream of messages, buffering can be turned off and a dedicated buffer per
connection can be used. When clients are expected to only sporadically send
messages, a pool can be used.